more efficient hashing for getLayerSeed

+++ src/BoundLayerable.sol
@@ -36,7 +36,7 @@ abstract contract BoundLayerable is RandomTraits, BoundLayerableEvents {
         string memory name,
         string memory symbol,
         address vrfCoordinatorAddress,
-        uint256 maxNumSets,
+        uint240 maxNumSets,
         uint8 numTokensPerSet,
         uint64 subscriptionId,
         ILayerable _metadataContract
